Fixed tests and one small bug.

+++ rv-predict-source/src/main/java/com/runtimeverification/rvpredict/util/Constants.java
@@ -16,7 +16,7 @@ public interface Constants {
 
     byte ATOMIC_LOCK_C = 43;
 
-    byte SIGNAL_LOCK_C = 44;
+    long SIGNAL_LOCK_C = 44;
 
     long INVALID_SIGNAL = -1L;
